Ignore conftest.py in code coverage
In addition, make it easier for produce coverage report locally by
adding pytest config options to pyproject.toml. This also produces
consistent results between local runs and runs in CI, and adds a
term-missing report for human readability, both for local runs as well
as in CI, so it show up in the logs.

Fixes #387