Add comment on skip_signature=True in the s3 store example
I am refactoring some of my code to virtualize NASA data to v2 at the moment, and just spent a bunch of minutes debugging a permission denied error due to me just copy/pasting the code in the example. 

Wondering if it would be instructive for users to know that this option is specific to the data used.